08_tracing/pyproject.toml
[project]
name = "tracing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentops>=0.4.16",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
]
